azure-certifications-study-guide/AI-projects/project-07-sentiment-analysis/src/requirements.txt
azure-functions==1.17.0
azure-ai-textanalytics==5.3.0
azure-cosmos==4.5.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
